[refactor] 폴더명 변경으로 인한 path 수정과 상태명 변경
